+++ Changes.md
@@ -17,7 +17,9 @@ Fixes
 
 - PlugAlgo : Updated `canSetValueFromData()`, `setValueFromData()` and `getValueAsData()` with support for missing types.
 - LightPosition Tool : Fixed lingering shadow pivot point after placing a shadow pivot, switching to highlight mode and switching back to shadow mode [^1].
-- Catalogue : Fixed bugs caused by reordering images using `GraphComponent::reorderChildren()`.
+- Catalogue :
+  - Fixed undo for image reordering via drag & drop.
+  - Fixed bugs caused by reordering images using `GraphComponent::reorderChildren()`.
 
 API
 ---